very lovely argc argv tasks

+++ 0x0A-argc_argv/2-args.c
@@ -1,14 +1,13 @@
 #include <stdio.h>
-#include "main.h"
 /**
- * main - func that prints the arguments received
- * @argc: Argument count
- * @argv: Argument vector
+ * main - Entry point
+ * @argc: argument count.
+ * @argv: argument vector.
  * Return: return (0) success
  */
 int main(int argc, char *argv[])
 {
-int i;
+unsigned short int i;
 for (i = 0; i < argc; i++)
 {
 printf("%s\n", argv[i]);
